pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-pptx-theme-kit"
version = "0.1.0"
description = "Reusable PowerPoint theming helpers built on python-pptx."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "python-pptx>=1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
